--- appveyor.yml
@@ -1,66 +1,0 @@
-environment:
-  next_version: 1.1.3
-  is_pr: false
-
-only_commits:
-  files:
-    - src/
-    - nuget/*.nuspec
-image: Visual Studio 2017
-configuration: Release
-init:
-  - ps: >-
-      if ($env:APPVEYOR_REPO_TAG -eq "true")
-      {
-        Update-AppveyorBuild -Version "$env:APPVEYOR_REPO_TAG_NAME"
-      }
-      else
-      {
-        if (-not $env:APPVEYOR_PULL_REQUEST_NUMBER)
-        {
-          $env:is_pr = "false";
-          if ($env:APPVEYOR_REPO_BRANCH -ne "master")
-          {
-            Update-AppveyorBuild -Version "$($env:next_version)-$($env:APPVEYOR_REPO_BRANCH)-build$($env:APPVEYOR_BUILD_NUMBER)"
-          }
-          else
-          {
-            Update-AppveyorBuild -Version "$($env:next_version)-beta$($env:APPVEYOR_BUILD_NUMBER)"
-          }
-        }
-        else
-        {
-          $env:is_pr = "true";
-          Update-AppveyorBuild -Version "$($env:next_version)-pr$($env:APPVEYOR_PULL_REQUEST_NUMBER)-build$($env:APPVEYOR_BUILD_NUMBER)"
-        }
-      }
-
-assembly_info:
-  patch: true
-  file: '**\AssemblyInfo.*'
-  assembly_version: '{version}'
-  assembly_file_version: '{version}'
-  assembly_informational_version: '{version}'
-before_build:
-- cmd: nuget restore src/LatestVersionPlugin.sln
-- cmd: msbuild src/Plugin.LatestVersion.macOS/Plugin.LatestVersion.macOS.csproj /verbosity:minimal
-build:
-  verbosity: minimal
-after_build:
-- cmd: nuget pack nuget/Plugin.LatestVersion.nuspec -basepath "./" -version "%APPVEYOR_BUILD_VERSION%"
-artifacts:
-- path: '**\*.nupkg'
-
-deploy:
-  - provider: NuGet
-    api_key:
-      secure: iZpf1YSkDiacJ9rKDwnS0DucuVDkzuEUhleT9T+ypZO7hhkToq+bmJCzuNyA9rkE
-    on:
-      branch: master
-      is_pr: false
-
-  - provider: NuGet
-    api_key:
-      secure: iZpf1YSkDiacJ9rKDwnS0DucuVDkzuEUhleT9T+ypZO7hhkToq+bmJCzuNyA9rkE
-    on:
-      appveyor_repo_tag: true
