--- nuget/Plugin.LatestVersion.nuspec
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
-    <metadata minClientVersion="2.8.1">
-        <id>Xam.Plugin.LatestVersion</id>
-        <title>LatestVersion Plugin for Xamarin and Windows</title>
-        <version>$version$</version>
-        <authors>edsnider</authors>
-        <owners>edsnider</owners>
-        <licenseUrl>https://github.com/edsnider/LatestVersionPlugin/blob/master/LICENSE</licenseUrl>
-        <projectUrl>https://github.com/edsnider/LatestVersionPlugin</projectUrl>
-        <iconUrl>https://raw.githubusercontent.com/edsnider/LatestVersionPlugin/master/art/icon.png</iconUrl>
-        <requireLicenseAcceptance>false</requireLicenseAcceptance>
-        <description>Easily detect if you are running the latest version of your Xamarin or UWP app and open it in the App Store, Play Store or Microsoft Store to update it.</description>
-        <tags>ios, android, macos, windows, xamarin.ios, xamarin.android, xamarin.mac, uwp, latestversion, apps, version, appstore, playstore, store, iphone, ipad, xamarin, xam.plugin, plugin</tags>
-        <releaseNotes>https://github.com/edsnider/LatestVersionPlugin/blob/master/CHANGELOG.md</releaseNotes>
-    </metadata>
-    <files>
-        <file src="src\Plugin.LatestVersion\bin\Release\netstandard1.1\Plugin.LatestVersion.*" target="lib\netstandard1.1" />
-        <file src="src\Plugin.LatestVersion.Android\bin\Release\Plugin.LatestVersion.*" target="lib\MonoAndroid10" />
-        <file src="src\Plugin.LatestVersion.iOS\bin\Release\Plugin.LatestVersion.*" target="lib\Xamarin.iOS10" />
-        <file src="src\Plugin.LatestVersion.macOS\bin\Release\Plugin.LatestVersion.*" target="lib\Xamarin.Mac20" />
-        <file src="src\Plugin.LatestVersion.UWP\bin\Release\Plugin.LatestVersion.*" target="lib\UAP10" />
-    </files>
-</package>
