+++ src/Exceptions.shared.cs
