+++ src/Extensions.uwp.cs
