+++ src/LatestVersion.uwp.cs
