--- src/Plugin.LatestVersion.Abstractions/Plugin.LatestVersion.Abstractions.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.1</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType></DebugType>
-    <DocumentationFile>bin\Release\netstandard1.1\Plugin.LatestVersion.Abstractions.xml</DocumentationFile>
-  </PropertyGroup>
-</Project>
