--- src/Plugin.LatestVersion.iOS/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("Plugin.LatestVersion")]
-[assembly: AssemblyProduct("Plugin.LatestVersion")]
-[assembly: AssemblyCopyright("2017 Ed Snider")]
-[assembly: AssemblyVersion("1.0.0")]
