--- src/Plugin.LatestVersion/Plugin.LatestVersion.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.1</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugType></DebugType>
-    <DocumentationFile>bin\Release\netstandard1.1\Plugin.LatestVersion.xml</DocumentationFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Plugin.LatestVersion.Abstractions\Plugin.LatestVersion.Abstractions.csproj" />
-  </ItemGroup>
-</Project>
