+++ src/Xam.Plugin.LatestVersion.csproj
@@ -1,0 +1,63 @@
+﻿<Project Sdk="MSBuild.Sdk.Extras/1.6.55">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard1.1;xamarin.ios10;xamarin.mac20;monoandroid71;uap10.0.14393</TargetFrameworks>
+
+    <RootNamespace>Plugin.LatestVersion</RootNamespace>
+    <AssemblyName>Plugin.LatestVersion</AssemblyName>
+    <AssemblyVersion>0.1.0.0</AssemblyVersion>
+    <AssemblyFileVersion>0.1.0.0</AssemblyFileVersion>
+
+    <!-- NuGet package info -->
+    <PackageId>Xam.Plugin.LatestVersion</PackageId>
+    <Title>LatestVersion Plugin for Xamarin and Windows</Title>
+    <PackageVersion>0.1.0</PackageVersion>
+    <Authors>edsnider</Authors>
+    <Owners>edsnider</Owners>
+    <PackageLicenseUrl>https://github.com/edsnider/LatestVersionPlugin/blob/master/LICENSE</PackageLicenseUrl>
+    <PackageProjectUrl>https://github.com/edsnider/LatestVersionPlugin</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/edsnider/LatestVersionPlugin</RepositoryUrl>
+    <PackageIconUrl>https://raw.githubusercontent.com/edsnider/LatestVersionPlugin/master/art/icon.png</PackageIconUrl>
+    <Description>Easily detect if you are running the latest version of your Xamarin or UWP app and open it in the App Store, Play Store or Microsoft Store to update it.</Description>
+    <PackageTags>ios, android, macos, windows, xamarin.ios, xamarin.android, xamarin.mac, uwp, latestversion, apps, version, appstore, playstore, store, iphone, ipad, xamarin, xam.plugin, plugin</PackageTags>
+    <PackageReleaseNotes>https://github.com/edsnider/LatestVersionPlugin/blob/master/CHANGELOG.md</PackageReleaseNotes>
+    
+    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
+    <DefineConstants>$(DefineConstants);</DefineConstants>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+    <DebugType>full</DebugType>
+    <DebugSymbols>true</DebugSymbols>
+    <GenerateDocumentationFile>false</GenerateDocumentationFile>
+  </PropertyGroup>
+
+  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
+    <DebugType>pdbonly</DebugType>
+    <GenerateDocumentationFile>true</GenerateDocumentationFile>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="**\*.shared.cs" />
+  </ItemGroup>
+  
+  <ItemGroup Condition=" $(TargetFramework.StartsWith('xamarin.ios')) ">
+		<Reference Include="System.Json" />
+    <Compile Include="**\*.ios.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" $(TargetFramework.StartsWith('xamarin.mac')) ">
+		<Reference Include="System.Json" />
+		<Compile Include="**\*.ios.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" $(TargetFramework.StartsWith('monoandroid')) ">
+    <Compile Include="**\*.android.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" $(TargetFramework.StartsWith('uap10.0')) ">
+    <Compile Include="**\*.uwp.cs" />
+  </ItemGroup>
+
+</Project>
