+++ src/Xam.Plugin.LatestVersion.sln
@@ -1,0 +1,16 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xam.Plugin.LatestVersion", "Xam.Plugin.LatestVersion.csproj", "{FED7F5AB-FEBB-41FD-A29E-325733A4C1E4}"
+EndProject
+Global
+    GlobalSection(SolutionConfigurationPlatforms) = preSolution
+        Debug|Any CPU = Debug|Any CPU
+        Release|Any CPU = Release|Any CPU
+    EndGlobalSection
+    GlobalSection(ProjectConfigurationPlatforms) = postSolution
+        {FED7F5AB-FEBB-41FD-A29E-325733A4C1E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+        {FED7F5AB-FEBB-41FD-A29E-325733A4C1E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+        {FED7F5AB-FEBB-41FD-A29E-325733A4C1E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+        {FED7F5AB-FEBB-41FD-A29E-325733A4C1E4}.Release|Any CPU.Build.0 = Release|Any CPU
+    EndGlobalSection
+EndGlobal
